+++ src/index.ts
@@ -8,8 +8,3 @@ export type { Plugin, Filter, Options, RequestHandler } from './types';
  * Default plugins
  */
 export * from './plugins/default';
-
-/**
- * Legacy exports
- */
-export * from './legacy';