+++ examples/express/index.js
@@ -1,8 +1,10 @@
 /**
  * Module dependencies.
  */
-const express = require('express');
-const { createProxyMiddleware } = require('../../dist'); // require('http-proxy-middleware');
+import express from 'express';
+import open from 'open';
+
+import { createProxyMiddleware } from '../../dist/index.js';
 
 /**
  * Configure proxy middleware
@@ -25,7 +27,7 @@ const server = app.listen(3000);
 console.log('[DEMO] Server: listening on port 3000');
 console.log('[DEMO] Opening: http://localhost:3000/users');
 
-require('open')('http://localhost:3000/users');
+open('http://localhost:3000/users');
 
 process.on('SIGINT', () => server.close());
 process.on('SIGTERM', () => server.close());