+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - fix(error-response-plugin): sanitize input
 - feat: drop node v14/v16/v18 [BREAKING CHANGE]
 - refactor(#1136): replace http-proxy w/ httpxy [BREAKING CHANGE]
+- chore: remove legacyCreateProxyMiddleware() [BREAKING CHANGE]
 
 ## [v3.0.5](https://github.com/chimurai/http-proxy-middleware/releases/tag/v3.0.5)