+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - fix(fixRequestBody): support content-encoding on request body (#1142)
 - fix: prevent TypeError when ws enabled but server is undefined (#1163)
 - fix: applyPathRewrite logs old req.url instead of rewritten path (#1157)
+- feat(hono): support for hono with createHonoProxyMiddleware
 
 ## [v3.0.5](https://github.com/chimurai/http-proxy-middleware/releases/tag/v3.0.5)