+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - feat(ipv6): support literal IPv6 addresses in `target` and `forward` options (ie. "http://[::1]:8000")
 - chore(package.json): bump httpxy to ^0.5.1
 - fix(logger-plugin): support ipv6 host and handle undefined protocol/host
+- ci(publish.yml): pin github.triggering_actor
 
 ## [v3.0.5](https://github.com/chimurai/http-proxy-middleware/releases/tag/v3.0.5)